+++ src/main/java/com/github/electroluxv2/utils/BackupScriptParameters.java
@@ -1,6 +1,6 @@
 package com.github.electroluxv2.utils;
 
-import com.github.electroluxv2.mixin.ServerWorldAccessor;
+import com.github.electroluxv2.mixin.ServerLevelAccessor;
 import net.minecraft.server.MinecraftServer;
 
 import java.util.ArrayList;
@@ -10,20 +10,21 @@ import java.util.stream.StreamSupport;
 public record BackupScriptParameters(
         List<String> levelNames,
         int playerCount,
-        float avgTickTime
+        float avgTickTime,
+        boolean running
 ) {
-    public static BackupScriptParameters fromServer(final MinecraftServer server) {
-        var playerCount = server.getCurrentPlayerCount();
-        var avgTickTime = server.getAverageTickTime();
+    public static BackupScriptParameters fromServer(final MinecraftServer server, boolean running) {
+        var playerCount = server.getPlayerCount();
+        var avgTickTime = server.getAverageTickTimeNanos();
 
         var levelNames = StreamSupport
-                .stream(server.getWorlds().spliterator(), false)
-                .map(world -> (ServerWorldAccessor) world.toServerWorld())
-                .map(world -> world.getWorldProperties().getLevelName())
+                .stream(server.getAllLevels().spliterator(), false)
+                .map(level -> (ServerLevelAccessor) level)
+                .map(level -> level.getServerLevelData().getLevelName())
                 .distinct()
                 .toList();
 
-        return new BackupScriptParameters(levelNames, playerCount, avgTickTime);
+        return new BackupScriptParameters(levelNames, playerCount, avgTickTime, running);
     }
 
     public List<String> toArguments() {
@@ -32,6 +33,8 @@ public record BackupScriptParameters(
         arguments.add(String.valueOf(playerCount));
 
         arguments.add(String.valueOf(avgTickTime));
+
+        arguments.add(String.valueOf(running));
 
         arguments.add(String.valueOf(levelNames.size()));
         arguments.addAll(levelNames);